requirements.txt
django==3.2.7
psycopg2
pillow
flask
django-ckeditor
pandas
pycryptodome